build: test 코드 작동에 맞춰 수정
