+++ .prettierignore
@@ -1,3 +1,3 @@
-src/main.legacy.js
-src/basic/**
-src/advanced/**
+src/main.original.js
+src/basic/**/__tests__/**
+src/advanced/**/__tests__/**
