+++ index.html
@@ -8,7 +8,7 @@
 </head>
 <body class="bg-gray-100 p-8">
 <main id="app"></main>
-<script type="module" src="/src/main.legacy.js"></script>
-<!--<script type="module" src="/src/main.original.js"></script>-->
+<script type="module" src="/src/main.original.js"></script>
+<script type="module" src="/src/basic/main.basic.js"></script>
 </body>
 </html>
